Simplify the code of exporting case runs
